refactor: Move filterExpression from web to packages
