+++ packages/filters/types.ts
@@ -8,3 +8,27 @@ export interface UIFilterRowNode {
   rows: UIFilterRowTree[];
 }
 export type UIFilterRowTree = UIFilterRowNode | UIFilterRow;
+
+// Re-export all types from filterExpressions
+export * from "./filterExpressions";
+export type {
+  FilterExpression,
+  AllExpression,
+  ConditionExpression,
+  AndExpression,
+  OrExpression,
+  AggregationExpression,
+  FilterOperator,
+  FieldSpec,
+  RequestResponseRMT,
+  FilterSubType,
+  Views,
+  UserMetric,
+} from "./filterExpressions";
+
+export {
+  FilterAST,
+  DEFAULT_FILTER_EXPRESSION,
+  DEFAULT_FILTER_GROUP_EXPRESSION,
+  EMPTY_FILTER_GROUP_EXPRESSION,
+} from "./filterExpressions";
