+++ web/filterAST/components/FilterGroupNode.tsx
@@ -7,7 +7,7 @@ import {
   DEFAULT_FILTER_GROUP_EXPRESSION,
   FilterAST,
   OrExpression,
-} from "../filterAst";
+} from "@helicone-package/filters";
 import { useFilterStore } from "../store/filterStore";
 import FilterConditionNode from "./FilterConditionNode";
 import { Row } from "@/components/layout/common/row";
