+++ web/filterAST/components/ui/SearchableSelect.tsx
@@ -16,7 +16,7 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from "@/components/ui/popover";
-import { FilterSubType } from "@/filterAST/filterAst";
+import { FilterSubType } from "@helicone-package/filters";
 
 export type SearchableSelectOption = {
   label: string;
