+++ web/filterAST/toFilterNode.ts
@@ -11,7 +11,7 @@ import {
   AndExpression,
   OrExpression,
   FieldSpec,
-} from "./filterAst";
+} from "@helicone-package/filters";
 
 /**
  * Maps the new FilterAST operator names to the legacy operator names
